updated CR deletion logic for when MCE goes away (#940)
* updated CR deletion logic for when MCE goes away


* added cluster-manager


* handled result in finalizeBackplaneConfig


* fixed the unit test bug (it was my bug)


* updated tests


* updated tests for k8sClient.Get()


* test line


* got the tests to pass


* tests are broken trying to throw errors


* tests are passing


* self code review


* updated error lines


---------

+++ controllers/backplaneconfig_controller.go
@@ -223,10 +223,14 @@ func (r *MultiClusterEngineReconciler) Reconcile(ctx context.Context, req ctrl.R
 	// If deletion detected, finalize backplane config
 	if backplaneConfig.GetDeletionTimestamp() != nil {
 		if controllerutil.ContainsFinalizer(backplaneConfig, backplaneFinalizer) {
-			err := r.finalizeBackplaneConfig(ctx, backplaneConfig) // returns all errors
+			result, err := r.finalizeBackplaneConfig(ctx, backplaneConfig) // returns all errors
 			if err != nil {
 				r.Log.Info(err.Error())
 				return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
+			}
+			r.Log.Info(fmt.Sprintf("Result returned from finalizeBackplaneConfig: %v", result))
+			if result != (ctrl.Result{}) {
+				return result, nil
 			}
 
 			r.Log.Info("all subcomponents have been finalized successfully - removing finalizer")
@@ -254,7 +258,7 @@ func (r *MultiClusterEngineReconciler) Reconcile(ctx context.Context, req ctrl.R
 		return ctrl.Result{}, err
 	}
 	if err != nil {
-		return ctrl.Result{}, err
+		return ctrl.Result{Requeue: true}, err
 	}
 
 	if !utils.ShouldIgnoreOCPVersion(backplaneConfig) && utils.DeployOnOCP() {
@@ -275,7 +279,7 @@ func (r *MultiClusterEngineReconciler) Reconcile(ctx context.Context, req ctrl.R
 		return ctrl.Result{}, err
 	}
 	if err != nil {
-		return ctrl.Result{}, err
+		return ctrl.Result{Requeue: true}, err
 	}
 
 	/*
@@ -295,7 +299,7 @@ func (r *MultiClusterEngineReconciler) Reconcile(ctx context.Context, req ctrl.R
 		return ctrl.Result{}, err
 	}
 	if err != nil {
-		return ctrl.Result{}, err
+		return ctrl.Result{Requeue: true}, err
 	}
 
 	// Attempt to retrieve image overrides from environmental variables.
@@ -807,6 +811,7 @@ func (r *MultiClusterEngineReconciler) ensureInternalEngineComponent(
 	backplaneConfig *backplanev1.MultiClusterEngine,
 	component string) (ctrl.Result, error) {
 
+	log.Info(fmt.Sprintf("Ensuring InternalEngineComponent: %v", component))
 	iec := &backplanev1.InternalEngineComponent{
 		TypeMeta: metav1.TypeMeta{
 			APIVersion: backplanev1.GroupVersion.String(),
@@ -840,22 +845,29 @@ func (r *MultiClusterEngineReconciler) ensureNoInternalEngineComponent(
 	ctx context.Context,
 	backplaneConfig *backplanev1.MultiClusterEngine,
 	component string) (ctrl.Result, error) {
-
+	log.Info(fmt.Sprintf("Ensuring No InternalEngineComponent: %s", component))
 	comp := &backplanev1.InternalEngineComponent{}
 	err := r.Client.Get(ctx, types.NamespacedName{Name: component, Namespace: backplaneConfig.Spec.TargetNamespace}, comp)
 	if apierrors.IsNotFound(err) {
+		log.Info(fmt.Sprintf("Component not found during Get(): %v", component))
 		return ctrl.Result{}, nil
 	}
 	if err != nil {
 		log.Error(err, fmt.Sprintf("Error getting InternalEngineComponent. Error: %v: Component was: %v", err, comp))
-		return reconcile.Result{}, err
+		return reconcile.Result{RequeueAfter: requeuePeriod}, err
 	}
 
-	log.Info(fmt.Sprintf("Ensuring No InternalEngineComponent: %s", component))
+	if comp.DeletionTimestamp != nil {
+		log.Info(fmt.Sprintf("DeletionTimestamp found: %v", comp.DeletionTimestamp))
+		log.Info("Skipping resource deletion")
+		return reconcile.Result{RequeueAfter: requeuePeriod}, nil
+	}
+
+	log.Info(fmt.Sprintf("Deleting %v InternalEngineComponent", component))
 	err = r.Client.Delete(ctx, comp)
 	if err != nil && !apierrors.IsNotFound(err) {
 		log.Error(err, fmt.Sprintf("Error deleting InternalEngineComponent. Error: %v", err))
-		return reconcile.Result{}, err
+		return reconcile.Result{RequeueAfter: requeuePeriod}, err
 	}
 
 	return ctrl.Result{}, nil
@@ -903,6 +915,7 @@ func (r *MultiClusterEngineReconciler) fetchChartOrCRDPath(component string, use
 
 func (r *MultiClusterEngineReconciler) ensureToggleableComponents(ctx context.Context,
 	backplaneConfig *backplanev1.MultiClusterEngine) (ctrl.Result, error) {
+	log.Info("Ensuring Toggleable Components")
 
 	errs := map[string]error{}
 	requeue := false
@@ -971,7 +984,7 @@ func (r *MultiClusterEngineReconciler) ensureToggleableComponents(ctx context.Co
 	if utils.DeployOnOCP() {
 		ocpConsole, err := r.CheckConsole(ctx)
 		if err != nil {
-			return ctrl.Result{RequeueAfter: requeuePeriod}, err
+			return ctrl.Result{}, err
 		}
 
 		if backplaneConfig.Enabled(backplanev1.ConsoleMCE) && ocpConsole {
@@ -986,6 +999,9 @@ func (r *MultiClusterEngineReconciler) ensureToggleableComponents(ctx context.Co
 			result, err = r.ensureNoConsoleMCE(ctx, backplaneConfig, ocpConsole)
 			if result != (ctrl.Result{}) {
 				requeue = true
+			}
+			if err != nil {
+				errs[backplanev1.ConsoleMCE] = err
 			}
 		}
 	}
@@ -1394,19 +1410,71 @@ func (r *MultiClusterEngineReconciler) ensureOpenShiftNamespaceLabel(ctx context
 	return ctrl.Result{}, nil
 }
 
-func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Context,
-	backplaneConfig *backplanev1.MultiClusterEngine) error {
+func (r *MultiClusterEngineReconciler) ensureNoAllInternalEngineComponents(ctx context.Context,
+	backplaneConfig *backplanev1.MultiClusterEngine) (ctrl.Result, error) {
+	errs := map[string]error{}
+	requeue := false
 
+	components := []string{
+		backplanev1.ConsoleMCE,
+		backplanev1.ManagedServiceAccount,
+		backplanev1.Discovery,
+		backplanev1.Hive,
+		backplanev1.AssistedService,
+		backplanev1.ServerFoundation,
+		backplanev1.ImageBasedInstallOperator,
+		backplanev1.ClusterLifecycle,
+		backplanev1.HyperShift,
+		backplanev1.ClusterProxyAddon,
+		backplanev1.LocalCluster,
+		backplanev1.ClusterManager,
+	}
+
+	for _, v := range components {
+		result, err := r.ensureNoInternalEngineComponent(ctx, backplaneConfig, v)
+		if result != (ctrl.Result{}) {
+			requeue = true
+		}
+		if err != nil {
+			errs[v] = err
+		}
+	}
+
+	if len(errs) > 0 {
+		errorMessages := []string{}
+		for k, v := range errs {
+			errorMessages = append(errorMessages, fmt.Sprintf("error ensuring %s: %s", k, v.Error()))
+		}
+		combinedError := fmt.Sprintf(": %s", strings.Join(errorMessages, "; "))
+		return ctrl.Result{}, errors.New(combinedError)
+	}
+	if requeue {
+		return ctrl.Result{RequeueAfter: requeuePeriod}, nil
+	}
+
+	return ctrl.Result{}, nil
+}
+
+func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Context,
+	backplaneConfig *backplanev1.MultiClusterEngine) (reconcile.Result, error) {
+
+	result, err := r.ensureNoAllInternalEngineComponents(ctx, backplaneConfig)
+	if err != nil {
+		return ctrl.Result{}, err
+	}
+	if result != (ctrl.Result{}) {
+		return result, nil
+	}
 	if utils.DeployOnOCP() {
 		ocpConsole, err := r.CheckConsole(ctx)
 		if err != nil {
-			return err
+			return ctrl.Result{}, err
 		}
 		if ocpConsole {
 			_, err := r.removePluginFromConsoleResource(ctx)
 			if err != nil {
 				log.Info("Error ensuring plugin is removed from console resource")
-				return err
+				return ctrl.Result{}, err
 			}
 		}
 	}
@@ -1414,7 +1482,7 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 	// Remove hypershift-addon ManagedClusterAddOn if present
 	hypershiftAddon, err := renderer.RenderHypershiftAddon(backplaneConfig)
 	if err != nil {
-		return err
+		return ctrl.Result{}, err
 	}
 	err = r.Client.Get(ctx, types.NamespacedName{Name: hypershiftAddon.GetName(),
 		Namespace: hypershiftAddon.GetNamespace()}, hypershiftAddon)
@@ -1422,14 +1490,14 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 	if err != nil {
 		if !(apierrors.IsNotFound(err) || apimeta.IsNoMatchError(err)) {
 			log.Error(err, "error while looking for hypershift-addon ManagedClusterAddOn")
-			return err
+			return ctrl.Result{}, err
 		}
 	} else {
 		log.Info("finalizing hypershift-addon ManagedClusterAddOn")
 		err := r.Client.Delete(ctx, hypershiftAddon)
 		if err != nil {
 			log.Error(err, "error deleting hypershift-addon ManagedClusterAddOn")
-			return err
+			return ctrl.Result{}, err
 		}
 
 		// If wait time exceeds expected then uninstall may not be able to progress
@@ -1441,7 +1509,7 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 			r.StatusManager.AddCondition(terminatingCondition)
 		}
 
-		return fmt.Errorf("waiting for 'hypershift-addon' ManagedClusterAddOn to be terminated before proceeding with uninstallation")
+		return ctrl.Result{}, fmt.Errorf("waiting for 'hypershift-addon' ManagedClusterAddOn to be terminated before proceeding with uninstallation")
 	}
 
 	localCluster := &unstructured.Unstructured{}
@@ -1458,7 +1526,7 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 		err := r.Client.Delete(ctx, localCluster)
 		if err != nil {
 			log.Error(err, "error deleting local-cluster ManagedCluster CR")
-			return err
+			return ctrl.Result{}, err
 		}
 
 		// If wait time exceeds expected then uninstall may not be able to progress
@@ -1476,12 +1544,12 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 
 			r.StatusManager.AddCondition(terminatingCondition)
 		}
-		return fmt.Errorf(
+		return ctrl.Result{}, fmt.Errorf(
 			"waiting for 'local-cluster' ManagedCluster to be terminated before proceeding with uninstallation")
 
 	} else if err != nil && !apierrors.IsNotFound(err) { // Return error, if error is not not found error
 		log.Error(err, "error while looking for local-cluster ManagedCluster CR")
-		return err
+		return ctrl.Result{}, err
 	}
 
 	clusterManager := &unstructured.Unstructured{}
@@ -1498,7 +1566,7 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 		log.Info("finalizing cluster-manager custom resource")
 		err := r.Client.Delete(ctx, clusterManager)
 		if err != nil {
-			return err
+			return ctrl.Result{}, err
 		}
 		// If wait time exceeds expected then uninstall may not be able to progress
 		if time.Since(backplaneConfig.DeletionTimestamp.Time) < 10*time.Minute {
@@ -1515,11 +1583,11 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 
 			r.StatusManager.AddCondition(terminatingCondition)
 		}
-		return fmt.Errorf(
+		return ctrl.Result{}, fmt.Errorf(
 			"waiting for 'cluster-manager' ClusterManager to be terminated before proceeding with uninstallation")
 
 	} else if err != nil && !apierrors.IsNotFound(err) { // Return error, if error is not not found error
-		return err
+		return ctrl.Result{}, err
 	}
 
 	ocmHubNamespace := &corev1.Namespace{}
@@ -1542,10 +1610,10 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 			r.StatusManager.AddCondition(terminatingCondition)
 		}
 
-		return fmt.Errorf(
+		return ctrl.Result{}, fmt.Errorf(
 			"waiting for 'open-cluster-management-hub' namespace to be terminated before proceeding with uninstallation")
 	} else if err != nil && !apierrors.IsNotFound(err) { // Return error, if error is not not found error
-		return err
+		return ctrl.Result{}, err
 	}
 
 	globalSetNamespace := &corev1.Namespace{}
@@ -1553,13 +1621,13 @@ func (r *MultiClusterEngineReconciler) finalizeBackplaneConfig(ctx context.Conte
 	if err == nil {
 		err := r.Client.Delete(ctx, globalSetNamespace)
 		if err != nil {
-			return err
+			return ctrl.Result{}, err
 		}
 	} else if err != nil && !apierrors.IsNotFound(err) {
-		return err
+		return ctrl.Result{}, err
 	}
 
-	return nil
+	return ctrl.Result{}, nil
 }
 
 func (r *MultiClusterEngineReconciler) getBackplaneConfig(ctx context.Context, req ctrl.Request) (
